// Web/src/app/models/visit-repair/visit-car-card.model.ts
export interface VisitCarCardModel {
    acceptanceDate: Date;
    receiptDate: Date;
    mileage: string;
    estimateCost: string;
    valuables: string;
    registrationDocument: boolean;
    ocDocument: boolean;
    keyLeft: boolean;
    testDriveConsent: boolean;
    usedPatrsReturn: boolean;
    faultDescription: string;
    externalConditionDescription: string;
}